Display page title and headling
This commit slightly refactors the plugin handling in order to separate
plugin instantiation from initialization. This allows each plugin to
have a complete list of all plugins, no matter the instantiation order.

Also the SPA router is changed to contain observable values to for the
overall site title and the current page name. Existing pages have been
adopted as a proof of concept. However the current pages will vanish
soon in favor of the really needed page types for course contents.