Replace old dummy pages with course content page
This commit removes the old dummy page types "home" and "user" in favor
of the really needed "course content" page type. Until now the course
content page type itself is also a placeholder which be fleshed out
soon.

Additionally a new config option called "home" has been added. It is
read by the new RedirectHomePlugin in order to redirect the user the
real home page if the root URL "/" is requested. This can e.g. be used
in order to use a special course as a welcome page.