Added translation capabilities
This commit adds the I18n plugin which can be used to translate the UI.
In order to create a new language, do the following:

  1. Create a new language file in src/locale, e.g. src/locale/fr.js
  2. Add the language in src/i18n.js to `availableLanguages` and
     `loadTranslation()`

A message catalog is a simple JavaScript module with one dictionary,
like this:

  "use strict";

  export default {
    "Translate this!": "Übersetze das!",
  }

And that's it. In order to access the translation do something like this
inside JavaScript:

  import plugins from "./app.js";
  const _ = plugins["I18n"].translate;
  ...

  var message = _("Translate this!");

In order to show a translated text on the UI bind it to an
ko.observable() just like any other computed value. Additionally the
current language and a dict of all available languages can be found in
the KO root context:

  * `$root.language`: The current language code
  * `$root.languages`: Dict with language codes and names

For anything else use the source, Luke!