Use hashbang URLs for development
Buzz-Light is a single page app using a client-site router for
navigation. In theory the SPA can be tested by setting the config option
devserver/historyApiFallback = true in webpack.js because then HTTP
request will return the index.html file. Without the setting you would
get a 404 instead when webpack auto-reloads the page after a file
modification. And while this somehow works in breaks source maps which
makes debugging a PITA.

The solution is to use hashbang URLs during development. Hashbang URLs
use the fact that the browser treats everything after a # sign in the
URL locally. Therefor no matter which page can be seen we always request
the index.html from the server. This allows us the get rid of the
historyApiFallback setting and enjoy source maps with proper testing of
the SPA.