Always reevaluate computed bindings which access translations
This is just a tiny refactoring which makes sure that computed bindings
which return translated messages will always be reevaluated when the
active language has been changed. The reason is that in order to display
a translated message in the UI the proper way is to use a computed
binding which calls plugins["I18n"].translate().