Display content of course pages
This commit adds to the core functionality of displaying course
contents. Content is defined in a so called manifest file named
"course.json" in the course's static file directory. The manifest
defines all pages and links to static HTML files.

HTML files can either be stand-alone or nested one level. If nested it
is possible to have different templates which determine how nested pages
are to be shown, e.g. as tab pages, guided steps and so on. Currently
however there are no such templates, yet. But the infrastructure is for
this is in place.

Each HTML file is treated like a knockout.js template and can thus use
the soon to be provided custom tags for their content.

BEWARE: Course contents are displayed using nested routers of the
ko-component-router library. But as it turns out this feature is way
buggy!!! So currently navigation within the page is broken until I find
a way to work around the library bugs ...