One SPA route to rule them all! First working implementation
Now try to find a decent single page app router on the net. Truth told:
There isn't. ko-component-router is a bag of bugs which have literally
eaten many days until I finally gave up with it. Then there is page.js
which seems to be unmaintained. Senna.js looks good at first sight but
it doesn't allow asynchronous route handlers.

But we need asynchronous route handlers. Why, I hear you ask? Because
at first we only know what courses there are but not which pages a
course contains. Only when the course manifest has been loaded and
parsed all valid URLs for that course are known. But of course we don't
want to pre-load all course manifests upfront but rather load them one
at a time when the user visits a course. Think of an installation with
thousands of courses here ...

So what we need to do is to first register some kind of a dummy route
like "^/courses/mycourse/*" which when called fetches the course data
and rewrites the routing rules. Turns out the existing libraries make
this extra hard for us. In the main branch this has basically been
solved but doesn't work right due to  ko-component-router's bugs.

So this is it. I give up with existing SPA router libraries. I mean, how
hard can it be? Really!? Call it Not Invented Here but rather than
fighting with someone else's half-assed code any longer I now enjoy
simply doing The Right Thing™ whenever I want to.