Adapted course screen to the new SPA router
With this commit the page is already working better than with
ko-componente-router as the single page app router. Granted the new
version takes more code, but then again the code is more structured and
the routing now seems to work okay.

FIXME: There is still a bug regarding backwards navigation. Instead of
pushing the current page on the history stack the moment it becomes
visible, one should push the page the moment it is left. Also at least
in Firefox the back button still does weird things and doesn't show
clear titles.