Debugged and fixed CourseScreen.onLeave() method
This is just a minor fix for the CourseScreen.onLeave() method. When the
user navigates to another course the specific routes for the previous
course are removed from the SPA router and replaced again with a generic
catch-all rule. However there was a tiny bug which led to the routes of
the new course being removed.