Breadcrumb navigational links added
This commit adds a breadcrumb below the current screen title. The
breadcrumb is only visible when a course page other than the start page
is shown. Additionally the provided example courses have been
restructured into a course overview and a tutorial in order to test the
new breadcrumb but also because this will be the default structure from
now on.