Added 500 screen where 404 is not sufficient
Previously when something went wrong most often the 404 screen was
shown. However the 404 screen is meant for situations where the
requested screen could not be found, this is where the URL doesn't map
to a valid routing rule. If there is however a routing rule but
something blows up while the screen is loaded now the 500 screen is
shown.