Fixing browser history
The HTML5 history API is a wild beast, as it turns out and though the
API itself is simple, it's hard to get it right. Who could know that
there is no way to distinguish between the user going back to the first
history entry and the user entering a new URL at the URL bar? In both
cases the popstate event is fired with a null state and no other
indicator. And who could know that setting the document title actually
modifies the current history entry but the title parameter of
history.pushState(state, title, url) and history.replaceState(state,
title, url) is ignored, instead. And who could know that setting a new
title doesn't modify the history immediately?

Turns out, I didn't. Let's see if I get this right the next time. For
know it seems to be working.