Resurrected the puzzled wizard from the fancy404 branch
This brings back the puzzled wizard on the 404 page. While still not
completely working as desired I think it is good enough to bring it back
to the main screen. When time permits I may fix the remaining bugs, but
at least it doesn't bit rot in its "The Deserted Git Branch" dungeon.