Use YAML instead of JSON for course manifest
Originally the course manifest was meant to be a JSON file because this
is a simple format for a later server-side extension of Buzz. However
the JSON format proved unhandy because it is very picky with trailing
commas (not allowed) and doesn't even allow for comments. So in order to
make the course manifest more human-readable YAML is the new format of
the day. And since JSON is a subset of YAML 1.2, JSON may still be used
where appropriate.