Fix URL on navigation and add <subpage-nav-buttons> HTML element
This commit adds a new <subpage-nav-buttons> custom HTML element which
can be used to show "Previous" and "Next" buttons to navigate between
subpages. Therefor one doesn't need to define one's own buttons with
hard links to the other pages.

Hand in hand with this addition goes a bug fix. The last commit which
made the SPA router remember scroll positions introduced a bug where the
URL in the browser's URL bar would change to an invalid value during
navigation to another screen.