Allow for white and black page backgrounds
This commit extends the course.yaml syntax with a "classes" property for
each page. This property can hold a list of CSS classes to be applied to
a page container, like this:

    ---
    language: ""
    name: "Buzz Tutorial"
    pages:
        "":
            name: "Start Page"
            file: "index.en.html"
            classes:
                - white
                - padding

The following classes are available:

  * white: white background with black font
  * black: black background with white font
  * padding: Adds a small padding to the page

Additionally the Buzz Tutorial has been extended to show off the new
capabilities.