First (non-responsive) version of chapters page type
This implements the "chapters" page type which is meant for a collection
of pages which belong together like the chapters of a book. Right now
the Table of Contents and the Page Content is displayed side by side but
I'm not entirely sure if I like this kind of display for as the page
content might grow much larger than the TOC. Also the layout is not
responsive, yet and thus only really works on a large screen.