Make fullscreen pages work correctly with Webkit and Chrome
This is just a minor fix which makes sure that fullscreen pages work
correctly with Webkit-based browsers even when their content is higher
than the viewport. The fix is simply to exchange this

  html, body {
      height: 100%;
  }

with this:

  html, body {
      height: 100vh;
  }

This seems to trigger another behaviour in Webkit regarding flexbox
content. Because fullscreen pages are shown inside a flexbox container
which vertically stretches a single item. However if the item's content
is larger than the initially calculated size, without the fix Webkit
overflows the content instead of making the item larger.